# docs/specs/openapi/control-plane.yaml
openapi: 3.1.0
info:
  title: Maestro Control Plane API
  version: 3.0.0
  license:
    name: Internal Proprietary
    identifier: LicenseRef-Internal-Proprietary
  description: >-
    Maestro v3 central control-plane contract. Project scope, role and delegated
    session are derived from the authenticated server context and resource path;
    request bodies never establish authorization scope. The API does not expose
    source-branch push, protected-branch push or merge operations.
servers:
  - url: /api/v3
security:
  - oidcBearer: []
...
paths:
...
  /projects/{project_id}:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      tags: [Projects]
      operationId: getProject
      summary: Read a Project snapshot
      description: Returns the authorized Project's current persisted snapshot.
      responses:
        '200':
          description: Project
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Update Project metadata or lifecycle state
      description: Applies an authorized optimistic update to Project metadata or state.
      x-maestro-permission: project.update
      x-maestro-audit-event: project.updated
      x-maestro-idempotency: required-header
      x-maestro-concurrency: if-match-required
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: {$ref: '#/components/schemas/ProjectUpdate'}
      responses:
        '200':
          description: Updated project
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
        '422': {$ref: '#/components/responses/Unprocessable'}
...
  /webhooks/dead-letters/{inbox_id}/replay:
    parameters:
      - {$ref: '#/components/parameters/InboxId'}
    post:
      tags: [GitLab]
      operationId: replayDeadLetter
      summary: Dual-approved replay of one quarantined webhook delivery
      description: >-
        Re-queues a dead-letter Inbox row under its ORIGINAL event identity:
        no re-verification bypass and no side-effect duplication. The approver
        identity is the authenticated principal; it must differ from
        requested_by (runbook section 9 dual-person control). The requeue, the
        approval and the audit row land in one transaction.
      x-maestro-permission: gitlab.reconcile
      x-maestro-audit-event: webhook.dead_letter.replayed
      x-maestro-idempotency: quarantined-row-state-collapse
      x-maestro-concurrency: single-row-transactional-guard
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DeadLetterReplay'}
      responses:
        '200':
          description: Quarantined row re-queued under its original identity
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeadLetterReplayResult'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookToken:
      type: apiKey
      in: header
      name: X-Gitlab-Token
  parameters:
...
    ProjectId:
      name: project_id
      in: path
      required: true
      description: Resource locator only; authorization scope is server-derived.
      schema: {$ref: '#/components/schemas/Identifier'}
...
    InboxId:
      name: inbox_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
...
  headers:
    ETag:
      description: Strong aggregate version ETag.
      schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
...
    Timestamp:
      type: string
      format: date-time
    ResourceVersion:
      type: integer
      minimum: 1
...
    ProjectState:
      type: string
      enum: [draft, configuring, active, suspended, archived]
    Project:
      type: object
      additionalProperties: false
      required: [id, team_id, key, name, state, version, created_at, updated_at]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        team_id: {$ref: '#/components/schemas/Identifier'}
        key: {type: string, pattern: '^[a-z][a-z0-9-]{1,31}$'}
        name: {type: string, minLength: 1, maxLength: 120}
        state: {$ref: '#/components/schemas/ProjectState'}
        version: {$ref: '#/components/schemas/ResourceVersion'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
...
    ProjectUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        state: {$ref: '#/components/schemas/ProjectState'}
...
    DeadLetterReplay:
      type: object
      additionalProperties: false
      required: [requested_by, reason]
      properties:
        requested_by:
          type: string
          minLength: 1
          maxLength: 200
          description: The principal who diagnosed and requested the replay.
        approved_by:
          type: string
          minLength: 1
          maxLength: 200
          description: >-
            Optional; when present it must name the authenticated principal —
            the server always derives the approver from the credential, never
            the body.
        reason:
          type: string
          minLength: 16
          maxLength: 2000
          description: Substantive justification recorded with the audit row.
    DeadLetterReplayResult:
      type: object
      additionalProperties: false
      required: [inbox_id, replayed]
      properties:
        inbox_id: {type: string, format: uuid}
        replayed: {type: boolean, const: true}
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlation_id]
      properties:
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]+$'}
        message: {type: string}
        correlation_id: {type: string}
        details: {type: object, additionalProperties: true}
  responses:
...
    BadRequest:
      description: Malformed request (`INVALID_REQUEST`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Authentication failed (`AUTHENTICATION_REQUIRED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: Authenticated principal lacks permission (`PERMISSION_DENIED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Resource absent or hidden by project visibility (`RESOURCE_NOT_FOUND`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: State, idempotency hash or uniqueness conflict (`STATE_CONFLICT`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    PreconditionFailed:
      description: ETag no longer matches (`VERSION_CONFLICT`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
...
    Unprocessable:
      description: Semantic validation or immutable rule failed (`VALIDATION_FAILED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    PreconditionRequired:
      description: Required conditional header missing (`PRECONDITION_REQUIRED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
...